requirements.txt
Babel
dj-database-url==0.5.0
Django
django-phonenumber-field
django-rest-framework
djangorestframework
gunicorn
phonenumbers
psycopg2
python-decouple
pytz
sqlparse
whitenoise